+++ src/components/transport_manager/include/transport_manager/android/android_connection_factory.h
@@ -1,4 +1,6 @@
 /*
+ * \file android_connection_factory.h
+ * \brief AndroidConnectionFactory class header file.
  *
  * Copyright (c) 2021, Ford Motor Company
  * All rights reserved.
@@ -31,45 +33,44 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "transport_manager/android_ipc/android_ipc_device.h"
-#include "transport_manager/common.h"
+#ifndef SRC_COMPONENTS_TRANSPORT_MANAGER_INCLUDE_TRANSPORT_MANAGER_ANDROID_ANDROID_CONNECTION_FACTORY_H_
+#define SRC_COMPONENTS_TRANSPORT_MANAGER_INCLUDE_TRANSPORT_MANAGER_ANDROID_ANDROID_CONNECTION_FACTORY_H_
 
-#include "utils/logger.h"
+#include "transport_manager/transport_adapter/server_connection_factory.h"
 
 namespace transport_manager {
 namespace transport_adapter {
-SDL_CREATE_LOG_VARIABLE("TransportManager")
 
-const ApplicationHandle kDefaultAppHandle = 1u;
+class AndroidTransportAdapter;
 
+/**
+ * @brief Create connections.
+ */
+class AndroidConnectionFactory : public ServerConnectionFactory {
+ public:
   /**
    * @brief Constructor.
    *
-   * @param device_address Bluetooth address.
-   * @param device_name Human-readable device name.
-   **/
-  AndroidIpcDevice::AndroidIpcDevice(const std::string& device_address,
-                  const char* device_name)
-    : Device(device_name, device_address)
-    , address_(device_address)
-    , applications_list_({kDefaultAppHandle}) { }
+   * @param controller Pointer to the device adapter controller.
+   */
+  AndroidConnectionFactory(AndroidTransportAdapter* controller);
 
-  bool AndroidIpcDevice::IsSameAs(const Device* other) const {
-      SDL_LOG_AUTO_TRACE();
-      const AndroidIpcDevice* other_bluetooth_device =
-      dynamic_cast<const AndroidIpcDevice*>(other);
-      if(other_bluetooth_device) {
-        return other_bluetooth_device->address_ == address_;
-      }
+ protected:
 
-      SDL_LOG_TRACE("Compare with no Android Ipc device !");
+  TransportAdapter::Error Init() override;
 
-      return false;
-  }
+  TransportAdapter::Error CreateConnection(
+      const DeviceUID& device_uid, const ApplicationHandle& app_handle) override;
 
-  ApplicationList AndroidIpcDevice::GetApplicationList() const {
-      return applications_list_;
-  }
+  void Terminate() override;
 
-}
-}
+  bool IsInitialised() const override;
+
+ private:
+  AndroidTransportAdapter* controller_;
+};
+
+}  // namespace transport_adapter
+}  // namespace transport_manager
+
+#endif  // SRC_COMPONENTS_TRANSPORT_MANAGER_INCLUDE_TRANSPORT_MANAGER_ANDROID_ANDROID_CONNECTION_FACTORY_H_
