+++ src/components/transport_manager/include/transport_manager/android/android_ipc_control_protocol.h
@@ -1,5 +1,5 @@
-#ifndef SRC_COMPONENTS_TRANSPORT_MANAGER_INCLUDE_TRANSPORT_MANAGER_ANDROID_IPC_ANDROID_IPC_CONTROL_PROTOCOL_H_
-#define SRC_COMPONENTS_TRANSPORT_MANAGER_INCLUDE_TRANSPORT_MANAGER_ANDROID_IPC_ANDROID_IPC_CONTROL_PROTOCOL_H_
+#ifndef SRC_COMPONENTS_TRANSPORT_MANAGER_INCLUDE_TRANSPORT_MANAGER_ANDROID_ANDROID_IPC_CONTROL_PROTOCOL_H_
+#define SRC_COMPONENTS_TRANSPORT_MANAGER_INCLUDE_TRANSPORT_MANAGER_ANDROID_ANDROID_IPC_CONTROL_PROTOCOL_H_
 
 #include <vector>
 #include <stdint.h>
@@ -32,4 +32,4 @@ struct AndroidIpcControlProtocol {
 }
 }
 
-#endif  // SRC_COMPONENTS_TRANSPORT_MANAGER_INCLUDE_TRANSPORT_MANAGER_ANDROID_IPC_ANDROID_IPC_CONTROL_PROTOCOL_H_
+#endif  // SRC_COMPONENTS_TRANSPORT_MANAGER_INCLUDE_TRANSPORT_MANAGER_ANDROID_ANDROID_IPC_CONTROL_PROTOCOL_H_
