--- src/components/transport_manager/include/transport_manager/android_ipc/ipc_receiver.h
@@ -1,30 +1,0 @@
-#ifndef SRC_COMPONENTS_TRANSPORT_MANAGER_INCLUDE_TRANSPORT_MANAGER_ANDROID_IPC_IPC_RECEIVER_H_
-#define SRC_COMPONENTS_TRANSPORT_MANAGER_INCLUDE_TRANSPORT_MANAGER_ANDROID_IPC_IPC_RECEIVER_H_
-/*#include <stdio.h>
-#include <stdlib.h>
-#include <unistd.h>
-#include <string.h>
-#include <sys/types.h>
-#include <sys/socket.h>
-#include <sys/un.h>
-#include <netinet/in.h>
-#include <netdb.h>*/
-
-#include "transport_manager/transport_adapter/transport_adapter_controller.h"
-
-namespace transport_manager {
-namespace transport_adapter {
-
-class IpcReceiver
-{
-public:
-    virtual ~IpcReceiver() {}
-    virtual void Init(const std::string& socket_name) = 0;
-    virtual void Run() = 0;
-    virtual void Stop() = 0;
-};
-
-}  // namespace transport_adapter
-}  // namespace transport_manager
-
-#endif  // SRC_COMPONENTS_TRANSPORT_MANAGER_INCLUDE_TRANSPORT_MANAGER_ANDROID_IPC_IPC_RECEIVER_H_
